Arviat Experimental DataSet
Find here a dataset for experimental testing. This dataset is from the Arviat area, a community situated on the western shore of Hudson Bay in Canada. The dataset comprises a MapExtent layer (in GeoPackage format), a Shoreline layer (in GeoPackage format), and a Digital Bathymetric Terrain Model (in GeoTiff format). Please note that the bathymetric data was collected by the Arviat community themselves.